Now you user can only delete the rating if its own or has admin role

+++ Backend/EdushareBackend/EdushareBackend/Controllers/RatingController.cs
@@ -34,9 +34,18 @@ namespace EdushareBackend.Controllers
         [Authorize]
         public void DeleteRating(string id)
         {
+            var userid = User.FindFirstValue(ClaimTypes.NameIdentifier);
+            var userRoleAdmin = User.IsInRole("Admin"); // vagy: User.IsInRole("Admin")
+
+            var uploaderId = ratingLogic.GetRatingById(id).UserId;
+
+            // ha nem admin ÉS nem a feltöltő → TILOS
+            if (!userRoleAdmin && userid != uploaderId)
+            {
+                throw new UnauthorizedAccessException("You are not allowed to delete this material");
+            }
+            
             ratingLogic.DeleteRatingById(id);
         }
-
-
     }
 }
